Replay of jmorgan29's PR IDAES/idaes-pse#18 here in idaes-dev
From @jmorgan29:
I found an issue with the code while running the "Tutorial_3_Dynamic_Flowsheets"
example. When plotting the component concentrations, the concentration of "NaOH"
is plotted for the curve labeled "Ethanol". Since it was a simple issue, I
created this PR in which the .py and .ipynb files are modified so that the
"Ethanol" concentration is plotted as suggested by the label.