Update pytest marks (#916)
* Rename `nocircleci` pytest mark to `integration`

* Mark slow tests.

This is an initial sweep of slow tests; some or even many may have
fallen through the cracks. Tests slower than 15s are marked with
@pytest.mark.integration; tests slower than 1s but faster than that are
marked with @pytest.mark.component.

* More slow test marks

Similar to 21b86fde61eb7af0eae9c13507180d7acc708fd8. I suspect that
there's something more complex going on with which tests are observed by
pytest to be slow (varying more than 8s seems unusual), so am moving on
with other changes before rooting out the details.

* Clean git merge detritus

* Add @pytest.mark.unit to all test functions...

... that are expected to take <1s. There are undoubtedly some slower
tests that have been included in the bunch, as well as tests marked with
@pytest.mark.slow.

* Add 'import pytest' to all test scripts.

* Add basic pytest marking script.

It currently lacks some usability such as having a commandline
interface. There is also a possible refactor incoming. Either way, it's
usable as-is.

* Mark a few test changes after merging

* Another missed commented test

* Yet more mark fixes for commented tests.

* More pytest import complications with multiline imports

Co-authored-by: Dan Gunter <dkgunter@lbl.gov>