Rename to main branch in docs
NOTE: this might result in awkward phrasing where "main" is used to refer to the upstream repository, e.g. "make sure your 'main' is up-to-date with the main 'main'".

We should address this once we have a preferred way to refer to the upstream repository (which could simply be "upstream").