Rename "master" to "main" (#46)
* Switch to main branch in CircleCI

* Rename to main branch in docs

NOTE: this might result in awkward phrasing where "main" is used to refer to the upstream repository, e.g. "make sure your 'main' is up-to-date with the main 'main'".

We should address this once we have a preferred way to refer to the upstream repository (which could simply be "upstream").

* Rename master_doc to main_doc

This is unrelated to the change in branch names, but it looked like a low-hanging (and hopefully non-breaking) fruit.