Adds configuration options to the column model
This adds fix_column_pressure and column_pressure configuration
options to the packed column model.  The fix_column_pressure dictates
whether the column pressure should be fixed at a given value or not, while
the column_pressure option is used to specific the desired value for the
column pressure. These options are used as conditions for adding
momemtum balance to the column model and they provide a convenient
means of maintaining the unit-model state before and after the
initialization rountine.