Corrected constraint expressions
Corrected the mathematical expressions for the constraints for ramping rates and startup/shutdown expressions. Namely, the expressions now appropriately constrain power ramp rates under all scenarios. Also, design decisions to build and it's impact on whether an equipment can turn on or not has been fixed. Minimum downtime constraint was also fixed. Also, I added a check to make sure that the up_time and down_time inputs are valid under the assumptions.