Updated cash flow construction
Updated the ability to make hourly cashflows. Now the user can provide all cashflows associated with all operation block models and the total cost and total revenue for each time point is generated to be used in the build cashflow function for the objective function. Still need to add tests for each of the logger messages (but the function appears to be working correctly).