Add capacity limits function
Added a function to add capacity limits for a given commodity. This was originally meant to be a user activity with linearization called at the operation_and_design_models level, however it appears it would be better in the price-taker-class itself. Discussions with Krishna have led us to use a NotImplementedError for a couple other things as well. All code now tested (except multi-year time horizons). Final testing offline for model format and finishing docstrings.