changes to PID model to get dynamic SCPP running
inital work changing PID controllers in the subcritical power plant (SCPP) model, this gets the dynamic simulation to run.
It appears the behavior of the calculate_initial_integral config was changed with 119ebb and this affected DOF for SCPP. switching a logic condition in controller.py allows for the SCPP dynamic simulation to run, but this should be revisited.

See TODO statements

also added an initialization test for dynamic SCPP